Move terminal init code to client_service module.
